Replace go-bindata with vfsgen
Fixes #49
Fixes #40

This removes a build time dependency that we previously need to install
in the environment. The frontend data can now be updated using "make
fronted_vfsdata.go". When running in "development mode" (that is using
the "dev" build tag, or through "make dev") the data file is ignored and
data read directly from disk, for easier iteration.