Remove ignore of JavaScript and JSON and CSS in downloads, expose as assets
Fixes #61

Introduce unfiltered assets, that can be freely inlcuded or made available for
download in the frontends: stop filtering downloads overly strict, and expose
all files that are not node documents or meta files as assets. Add `assets` to
API responses for nodes and deprecate `downloads`, it will be removed in APIv3.